Historical/requirements.txt
pandas==2.2.0
pyarrow==15.0.0
python-binance==1.0.19
SQLAlchemy==2.0.25